feat(FX-4607): Add "Artsy Editorial" prefix in activity panel
