+++ src/app/Scenes/Activity/utils/shouldDisplayNotificationTypeLabel.ts
@@ -1,0 +1,3 @@
+export const shouldDisplayNotificationTypeLabel = (notificationType: string) => {
+  return ["ARTWORK_ALERT", "ARTICLE_FEATURED_ARTIST"].includes(notificationType)
+}
